Plugin/Module Support

+++ database.go
@@ -2,23 +2,25 @@ package beubo
 
 import (
 	"fmt"
-	"github.com/jinzhu/gorm"
+	"github.com/uberswe/beubo/pkg/plugin"
 	"github.com/uberswe/beubo/pkg/utility"
+	"gorm.io/gorm"
 	"io/ioutil"
 	"log"
 
-	// Gorm recommends a blank import to support underlying mysql
-	_ "github.com/jinzhu/gorm/dialects/mysql"
-	_ "github.com/jinzhu/gorm/dialects/sqlite"
 	"github.com/uberswe/beubo/pkg/structs"
 	"golang.org/x/crypto/bcrypt"
+	// Gorm recommends a blank import to support underlying mysql
+	"gorm.io/driver/mysql"
+	"gorm.io/driver/sqlite"
 )
 
 var (
 	seedEmail    = "seed@beubo.com"
 	seedPassword = "Beubo1234!"
 	// TODO change this to a config
-	shouldSeed = true
+	shouldSeed            = false
+	shouldRefreshDatabase = false
 	// DB is used to perform database queries globally. In the future this should probably
 	// be changed so that database.go declares methods that can be used to perform types of
 	// queries
@@ -27,37 +29,45 @@ var (
 
 func setupDB() *gorm.DB {
 	log.Println("Opening database")
-	connectString := ""
-	if databaseDriver == "sqlite3" {
-		connectString = databaseName
-	} else {
-		connectString = fmt.Sprintf("%s:%s@tcp(%s:%s)/%s?charset=utf8&parseTime=True&loc=Local", databaseUser, databasePassword, databaseHost, databasePort, databaseName)
-	}
-	db, err := gorm.Open(databaseDriver, connectString)
+	dialector := getDialector(databaseUser, databasePassword, databaseHost, databasePort, databaseName, databaseDriver)
+	db, err := gorm.Open(dialector, &gorm.Config{})
 	utility.ErrorHandler(err, true)
 
 	return db
 }
 
+func getDialector(user string, pass string, host string, port string, name string, driver string) gorm.Dialector {
+	connectString := fmt.Sprintf("%s:%s@tcp(%s:%s)/%s?charset=utf8&parseTime=True&loc=Local", user, pass, host, port, name)
+	dialector := mysql.Open(connectString)
+	if driver == "sqlite3" {
+		connectString = databaseName
+		dialector = sqlite.Open(connectString)
+	}
+	return dialector
+}
+
 func databaseInit() {
 	DB = setupDB()
 
-	type Result struct {
-		DropQuery string
-	}
+	if shouldRefreshDatabase {
 
-	var result []Result
+		type Result struct {
+			DropQuery string
+		}
 
-	log.Println("Dropping all database tables")
+		var result []Result
 
-	if databaseDriver == "sqlite3" {
-		DB.Raw("SELECT 'DROP TABLE IF EXISTS `' || name || '`;'  as drop_query FROM sqlite_master WHERE type='table' AND name NOT LIKE 'sqlite_%';").Scan(&result)
-	} else {
-		DB.Raw("SELECT concat('DROP TABLE IF EXISTS `', table_name, '`;') as drop_query FROM information_schema.tables WHERE table_schema = 'beubo';").Scan(&result)
-	}
+		log.Println("Dropping all database tables")
 
-	for _, r := range result {
-		DB.Exec(r.DropQuery)
+		if databaseDriver == "sqlite3" {
+			DB.Raw("SELECT 'DROP TABLE IF EXISTS `' || name || '`;'  as drop_query FROM sqlite_master WHERE type='table' AND name NOT LIKE 'sqlite_%';").Scan(&result)
+		} else {
+			DB.Raw("SELECT concat('DROP TABLE IF EXISTS `', table_name, '`;') as drop_query FROM information_schema.tables WHERE table_schema = 'beubo';").Scan(&result)
+		}
+
+		for _, r := range result {
+			DB.Exec(r.DropQuery)
+		}
 	}
 
 	log.Println("Running database migrations")
@@ -73,7 +83,8 @@ func databaseInit() {
 		&structs.Site{},
 		&structs.Tag{},
 		&structs.Comment{},
-		&structs.Setting{})
+		&structs.Setting{},
+		&plugin.PluginSite{})
 }
 
 func prepareSeed(email string, password string) {
@@ -93,8 +104,10 @@ func databaseSeed() {
 	for _, file := range files {
 		// Ignore the install directory, only used for installation
 		if file.IsDir() && file.Name() != "install" {
-			theme = structs.Theme{Slug: file.Name(), Title: file.Name()}
-			if DB.NewRecord(theme) { // => returns `true` as primary key is blank
+			theme = structs.Theme{}
+			DB.Where("slug = ?", file.Name()).First(&theme)
+			if theme.ID == 0 {
+				theme = structs.Theme{Slug: file.Name(), Title: file.Name()}
 				DB.Create(&theme)
 			}
 		}
@@ -110,15 +123,17 @@ func databaseSeed() {
 		utility.ErrorHandler(err, true)
 
 		user := structs.User{Email: testuser, Password: string(hashedPassword)}
-
-		if DB.NewRecord(user) { // => returns `true` as primary key is blank
+		DB.Where("email = ?", user.Email).First(&user)
+		if user.ID == 0 {
 			DB.Create(&user)
 		}
+
 	}
 
 	// user registration is disabled by default
 	disableRegistration := structs.Setting{Key: "enable_user_registration", Value: "false"}
-	if DB.NewRecord(disableRegistration) { // => returns `true` as primary key is blank
+	DB.Where("key = ?", disableRegistration.Key).First(&disableRegistration)
+	if disableRegistration.ID == 0 {
 		DB.Create(&disableRegistration)
 	}
 
@@ -134,9 +149,7 @@ func databaseSeed() {
 
 		user := structs.User{Email: seedEmail, Password: string(hashedPassword)}
 
-		if DB.NewRecord(user) { // => returns `true` as primary key is blank
-			DB.Create(&user)
-		}
+		DB.Create(&user)
 
 		// Create a site
 
@@ -147,9 +160,7 @@ func databaseSeed() {
 			Type:   1,
 		}
 
-		if DB.NewRecord(site) {
-			DB.Create(&site)
-		}
+		DB.Create(&site)
 
 		// Create a page
 		content := `<p>Welcome to Beubo! Beubo is a free, simple, and minimal CMS with unlimited extensibility using plugins. This is the default page and can be changed in the admin area for this site.</p>`
@@ -166,9 +177,7 @@ func databaseSeed() {
 			SiteID:   int(site.ID),
 		}
 
-		if DB.NewRecord(page) {
-			DB.Create(&page)
-		}
+		DB.Create(&page)
 
 		shouldSeed = false
 		seedEmail = ""
